Model: Now the tower shots on the mobs. The mobs can be deletet from the field. But there some problems with the deletion from the mob. So we must look at the problem.

+++ TowerDefence/src/de/htwg/towerdefence/model/impl/Mob.java
@@ -133,10 +133,12 @@ public class Mob extends ControllableComponent implements IMob {
 			gameContext.getCheckWay().existWay(this.currendPos.getX() , this.currendPos.getY(), 9, 9);
 			List<Coord> way = gameContext.getCheckWay().getShortesWay();
 			
-			gameContext.getPlayingField().deleteMob(this.currendPos.getX(), this.currendPos.getY(), this);
-			gameContext.getPlayingField().setMob(way.get(1).getX(), way.get(1).getY(), this);
-			this.currendPos.setX(way.get(1).getX());
-			this.currendPos.setY(way.get(1).getY());
+			if (gameContext.getPlayingField().deleteMob(this.currendPos.getX(), this.currendPos.getY(), this) != null){
+				gameContext.getPlayingField().deleteMob(this.currendPos.getX(), this.currendPos.getY(), this);
+				gameContext.getPlayingField().setMob(way.get(1).getX(), way.get(1).getY(), this);
+				this.currendPos.setX(way.get(1).getX());
+				this.currendPos.setY(way.get(1).getY());
+			}
 			
 			this.tmpSpeed = this.speed;
 			return true;
